@richardstartin 2017-05-08
Addressing  #139