@lemire 2017-05-08
@richardstartin 

I think that this is good. 

I left a tiny comment. I don't think it is worth doing anything about it, so I will merge your code.

Thanks @Kineolyan